+++ mudz.Core/Model/Domain/Player/Inventory/Item/Weapon/TestHammer.cs
@@ -7,10 +7,10 @@ namespace mudz.Core.Model.Domain.Player.Inventory.Item.Weapon
     {
         public TestHammer()
         {
-            ActionEffect = new Dictionary<InventoryAugmentEffect, double> { { InventoryAugmentEffect.Attack, 2 } };
+            ActionEffect = new Dictionary<InventoryAugmentEffect, int> { { InventoryAugmentEffect.Attack, 2 } };
         }
 
-        public override sealed Dictionary<InventoryAugmentEffect, double> ActionEffect { get; set; }
+        public override sealed Dictionary<InventoryAugmentEffect, int> ActionEffect { get; set; }
 
     }
 }
