+++ mudz.Core/Model/Domain/Player/Inventory/Item/Wearable/TestGloves.cs
@@ -7,10 +7,10 @@ namespace mudz.Core.Model.Domain.Player.Inventory.Item.Wearable
     {
         public TestGloves()
         {
-            ActionEffect = new Dictionary<InventoryAugmentEffect, double> { { InventoryAugmentEffect.Attack, 1 } };
+            ActionEffect = new Dictionary<InventoryAugmentEffect, int> { { InventoryAugmentEffect.Attack, 1 } };
         }
 
-        public override sealed Dictionary<InventoryAugmentEffect, double> ActionEffect { get; set; }
+        public override sealed Dictionary<InventoryAugmentEffect, int> ActionEffect { get; set; }
 
         public override PlayerAnatomy Anatomy {get { return PlayerAnatomy.Hands; }}
